task-backend/requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
python-multipart==0.0.6

sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.0

httpx==0.24.1
pydantic==1.10.12
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0